+++ Configuration/TSconfig/GridElements/ColumnLayoutGridElement.ts
@@ -1,3 +1,4 @@
 # === Includes ===
 
 <INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayoutGridElement/2Columns.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayoutGridElement/3Columns.ts">