+++ Configuration/TypoScript/GridElements/ColumnLayoutGridElement/constants.ts
@@ -1,3 +1,4 @@
 # === Includes ===
 
 <INCLUDE_TYPOSCRIPT: source="FILE: ./2Columns/constants.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./3Columns/constants.ts">