+++ Configuration/TypoScript/GridElements/ColumnLayoutGridElement/setup.ts
@@ -1,3 +1,4 @@
 # === Includes ===
 
 <INCLUDE_TYPOSCRIPT: source="FILE: ./2Columns/setup.ts">
+<INCLUDE_TYPOSCRIPT: source="FILE: ./3Columns/setup.ts">